+++ src/pages/Solicitudes/Nutricion/Nutricion.jsx
@@ -1,19 +1,19 @@
 import React from "react";
-import "./homepage.css";
-import logo from "./assets/logo-ucchristus.png";
+import "../../homepage.css";
+import logo from "../../../assets/logo-ucchristus.png";
 import { Link } from "react-router-dom";
 
-export default function AcompañamientoEspiritual() {
+export default function NutricionAlimentacion() {
   return (
     <div className="app">
       <img src={logo} alt="Logo UC Christus" className="logo" />
       <p>Por favor indíquenos de qué área es su consulta</p>
       <section className="botones">
-          <div className="bot botones_morados titulo-estatica"> ACOMPAÑAMIENTO SOCIAL Y/O ESPIRITUAL </div>
-          <Link className="bot botones_blancos" to="/acompañamiento_espiritual">
+          <div className="bot botones_morados titulo-estatica"> NUTRICIÓN Y ALIMENTACIÓN A PACIENTES </div>
+          <Link className="bot botones_blancos" to="/nutricion_y_alimentacion">
             PREGUNTAR A CONI
           </Link>
-          <Link className="bot botones_blancos" to="/acompañamiento_espiritual">
+          <Link className="bot botones_blancos" to="/nutricion_y_alimentacion">
             PREGUNTAR A CONI
           </Link>
           <Link className="bot botones_azules" to="/">
